Reorganise the repository
This change reorganises the repository so that the structure is as
simple to run and as intuitive as possible, given the nature of the
code. It also updates the README to reflect that change, along with
integrating instructions showing how to start, check, and stop the
application using Docker Compose, and how to run the code in each
section.